Fix argument typo in log message
It should be '%s' instead of '%'. It looks like Python is intepreting
the i from the next word as the argument instead, and so failing with a
TypeError:

--- Logging error ---
Traceback (most recent call last):
  File "/usr/lib64/python3.7/logging/__init__.py", line 1025, in emit
    msg = self.format(record)
  File "/usr/lib64/python3.7/logging/__init__.py", line 869, in format
    return fmt.format(record)
  File "/usr/lib64/python3.7/logging/__init__.py", line 608, in format
    record.message = record.getMessage()
  File "/usr/lib64/python3.7/logging/__init__.py", line 369, in getMessage
    msg = msg % self.args
TypeError: %i format: a number is required, not str
Call stack:
[...]
  File "/home/jpichon/devel/vmware-nsxlib/vmware_nsxlib/v3/policy/core_defs.py", line 478, in _version_dependant_attr_supported
    attr, self.nsx_version, nsx_constants.NSX_VERSION_3_0_0)
Message: 'Attribute % is not supported. Current NSX version %s, minimum supported version %s'
Arguments: ('pool_allocation', '0.0.0', '3.0.0')

Change-Id: I3c50571ad0f797ae2c35c28ad3e9c28a07881f91