Expose HTTP Status Code in ManagerError
When ManagerError is raised due to unavailability of NSX backend, it is
common that return HTTP status is non-200 and no error code is provided
by NSX. In this case, recipient of ManagerError can hardly check the
cause due to not having access to either error code or HTTP status code.

This patch includes the HTTP status code in ManagerError, aiming to
provide recipient a more reliable way to check the cause of error.

Change-Id: Id346b27ad69ca4c3b3b4ef950a8c191ce46cde5e