Allow disabling keepalive_section
In case validate_connection_method already has the effect to keep alive,
it should be allowed to not perform any extra keep-alive requests.

Currently in MP the default keepalive section is transport-zones, which
is deemed to degrade in performance a lot in scale setup. As a more
light-weighted path reverse-proxy/node/health is already used, we should
allow configuring keepalive section to be disabled.

Change-Id: I26c0af67f90b62533a39827ca5111832d306a153