Use reverse-proxy/node/health for connection status check
Previously nsxlib uses operation/application/status for nsx
connection status check if nsx version is <= 2.3. This endpoint
is known to be of high overhead in scale environments.
Furthermore, since self.nsx_version is not initialized during
the initialization phase, the validate_connection_method tend
to always use the v1 version of check.

Since nsx 2.3 is now out of the support matrix, nsxlib can simply
use reverse-proxy/node/health endpoint to verify connection to
nsx is up, which is very lightweighted.

Change-Id: Ife48d5c357e8afe416e2ad772e5bbb0616089b36
Signed-off-by: Yang Ding <dingyang@vmware.com>