Translate exception that grounds cluster
When endpoint goes down, the user should see same exception as
when the cluster is already down (detected by earlier activity).
For this purpose, translate grounding exception to
ServiceClusterUnavaliable.
In addition, display a warning if amount of retries is less than
amount of endpoints, since in this case not all endpoints will be
probed.

Change-Id: Ib4aa5eb95069b917c989b1f6dcd3535880b5a038