Add support for enabling/disabling multicast on Tier-1 GW
This patch allow for enabling or disabling multicast on a Tier-1 GW
router. This is simply done by setting a boolean flag in the router's
multicast configuration, which is however a sub-attribute of the
Tier-1 GW object accessed via its own API endpoint.

For the above reason, this patch introduces a definition object for
the Tier-1 Multicast settings.

Change-Id: I8308442ecd9b4d14f4ceb0ea55c4dcd4ee240e17